Fix stroke issues and allow floating point values

+++ src/animatable/element/text_element.hpp
@@ -19,7 +19,7 @@ class TextLayoutItem {
 
     int height;
     double opacity{1};
-    int strokeWidth{0};
+    double strokeWidth{0};
 };
 
 class TextLayout {
@@ -88,7 +88,7 @@ class TextAnimatorRender : public AnimatableRender {
     PropertyRender<double> y{this};
     PropertyRender<double> opacity{this};
     PropertyRender<int> letterSpacing{this};
-    PropertyRender<int> strokeWidth{this};
+    PropertyRender<double> strokeWidth{this};
 };
 
 class TextAnimator : public Animatable, public ICollapsible {
@@ -115,7 +115,7 @@ class TextAnimator : public Animatable, public ICollapsible {
     Property<double> y{this, "y", 0};
     Property<double> opacity{this, "opacity", 100};
     Property<int> letterSpacing{this, "letterSpacing", 0};
-    Property<int> strokeWidth{this, "strokeWidth", 0};
+    Property<double> strokeWidth{this, "strokeWidth", 0};
 
     void _propertyUpdated(PropertyBase *property) override;
     void _propertyIsAnimatingUpdated(PropertyBase *property) override;
